+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ### Unreleased
 
+* Rename `Entity.delinkFromInternalFilteringSubscription(Subscription)` method to `Entity.delinkFromFilteringSubscription(Subscription)` and modify the implementation to support calling with any filtered graph. The intent is to support mutable routing parameters.
+
 ### [v6.112](https://github.com/replicant4j/replicant/tree/v6.112) (2021-09-21) · [Full Changelog](https://github.com/spritz/spritz/compare/v6.111...v6.112)
 
 Changes in this release: